+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 
 ## Unreleased
 
+## v1.7.0 (2022-04-11)
+
 - Update dependencies
 - Add support for custom dimensions
-- Change form sequence to varagrs API in settings
+- Change from sequence to varagrs API in settings
 
 ## v1.6.0 (2021-05-07)